+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  roots: ['app/javascript'],
+  setupTestFrameworkScriptFile: './app/javascript/src/setupTests.js',
+};